requirements.txt
click==8.1.3
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2-binary==2.9.6
Pygments==2.14.0
rich==13.3.3
typer==0.7.0
